Added error checking to staff data migration to check for pages that have been edited.

+++ staff/migrations/0003_load_initial_staff_data.py
@@ -210,6 +210,9 @@ def create_staffpages(apps, schema_editor):
 def remove_staffpages(apps, schema_editor):
     StaffPage = apps.get_model('staff.StaffPage')
     for s in StaffPage.objects.all():
+        # check to see if someone edited this page. If they did, throw an error. 
+        if not s.first_published_at is None:
+            raise ValueError
         s.delete()
 
 class Migration(migrations.Migration):
